refactor(screenshots.yml): modify GitHub action to trigger on push instead of pull request feat(screenshots.yml): add paths and branches-ignore to limit action triggers fix(screenshots.yml): adjust concurrency group to include github.ref for uniqueness refactor(screenshots.yml): remove unnecessary conditions and fetch-depth in checkout action style(screenshots.yml): rename screenshot step to take-screenshot for clarity fix(screenshots.json): adjust geometry values for better screenshot positioning
The GitHub action is now triggered on push events instead of pull requests, which allows for more immediate feedback. The action is also limited to only trigger when changes are made to the screenshots.json file and not on main, dependabot, or release-please branches. The concurrency group now includes the github.ref to ensure uniqueness and prevent potential conflicts. Unnecessary conditions and fetch-depth in the checkout action were removed to simplify the workflow. The screenshot step was renamed to take-screenshot for better clarity. Lastly, the geometry values in screenshots.json were adjusted to improve the positioning of the screenshots.