fix(terraform.yml): modify path for downloaded artifact to correct directory feat(mkdocs.yml): add azure-service-principal.md to CI/CD section in navigation
The path for the downloaded artifact in the terraform.yml file was incorrect, causing the Terraform Apply step to fail. This change corrects the path, ensuring the artifact is placed in the correct directory for the apply step to succeed.

In the mkdocs.yml file, a new document azure-service-principal.md has been added to the CI/CD section in the navigation. This provides additional information on Azure Service Principal, enhancing the documentation's coverage of CI/CD topics.