docs(azure-service-principal.md): update command and instructions for setting up Azure service principal
The command for creating an Azure service principal has been updated to remove the --name flag, as it's not necessary and can lead to confusion. The instructions for setting the DEPLOYED variable have also been updated to clarify that it should be created manually using the Github UI. This change is made to avoid any confusion and to ensure that the setup process is as straightforward as possible.