chore(mkdocs-build-test.yml): ignore certain branches in workflow trigger feat(mkdocs-build-test.yml): add github.ref to concurrency group for unique identification
The workflow trigger has been updated to ignore certain branches such as 'gh-pages', 'main', 'dependabot*', and 'release-please*'. This is to prevent unnecessary workflow runs on these branches which are not intended for documentation updates.

The concurrency group identifier has been updated to include the github.ref. This allows for unique identification of workflow runs, preventing potential conflicts when multiple workflows are triggered simultaneously.